Revert "tour + refactoring"
This reverts commit 49ee18dc9900df93e080e29c48e79b7174ca6a73.
