+++ src/components/Modals/NamePlanModal.vue
@@ -78,7 +78,8 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '@/components/Modals/MultiplePlans/PlanModalDropdown.scss';
+@import '@/components/Modals/PlanModalDropdown.scss';
+
 .content-plan {
   width: 20rem;
 }
