+++ src/components/Modals/TextInputModal.vue
@@ -48,7 +48,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss">
-@import '@/components/Modals/MultiplePlans/PlanModalDropdown.scss';
+@import '@/components/Modals/PlanModalDropdown.scss';
 
 .textInput {
   &-label {
